fix(recovery): add AuthCheck for mnemonic
